# Conflicts:
#	app/src/main/java/com/tripleThreads/taxiyaz/data/location/Location.kt
#	app/src/main/java/com/tripleThreads/taxiyaz/fragments/AlternativeRoutingFragment.kt
#	app/src/main/java/com/tripleThreads/taxiyaz/fragments/add_route/BottomDialogFragment.kt

+++ app/src/main/java/com/tripleThreads/taxiyaz/data/comment/CommentDao.kt
@@ -8,7 +8,7 @@ import com.tripleThreads.taxiyaz.data.route.Route
 @Dao
 interface CommentDao {
     @Query("SELECT * FROM Comments where route_id=:routeId ORDER BY date")
-    fun getAllComments(routeId: Int): LiveData<List<Comment>>
+    fun getAllComments(routeId: Long): LiveData<List<Comment>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun insertComment(comment: Comment)
